archer-line-agent/pyproject.toml
[project]
name = "archer-line-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "DeanSCND-Symend",email = "dskelton@symend.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "cartesia-line (>=0.1.3,<0.2.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.104.0,<0.105.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
